postman testing was successful

+++ src/utils/cloudinary.js
@@ -5,20 +5,24 @@ import fs from "fs"
 cloudinary.config({ 
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
   api_key: process.env.CLOUDINARY_API_KEY, 
-  api_secret: process.env.CLOUDINARY_APL_SECRET 
+  api_secret: process.env.CLOUDINARY_APL_SECRET,
 });
 
-const uploadOnCloudinary =async(localFilePath)=>{
+const uploadOnCloudinary = async(localFilePath)=>{
   try{
     if(!localFilePath) return null  
     const response = await cloudinary.uploader.upload(localFilePath,{
       resource_type:"auto"
-    })
-    console.log("file is uploaded on cloudinary",response.url)
+    });
+    // console.log("file is uploaded on cloudinary",response.url)
+    fs.unlinkSync(localFilePath)
     return response;
 
   } catch(error){
-    fs.unlinkSync(localFilePath)
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
+    // fs.unlinkSync(localFilePath)
     return null;
   }
 }
